+++ CHANGES.md
@@ -23,6 +23,7 @@ Release Notes.
 * Provide HikariCP database connection pool plugin.
 * Fix NumberFormat exception in jdbc-commons plugin when MysqlURLParser parser jdbcurl
 * Provide Alibaba Fastjson parser/generator plugin.
+* Provide Jackson serialization and deserialization plugin.
 * Fix a tracing context leak of SpringMVC plugin, when an internal exception throws due to response can't be found.
 * Make GRPC log reporter sharing GRPC channel with other reporters of agent. Remove config items of `agent.conf`, `plugin.toolkit.log.grpc.reporter.server_host`, `plugin.toolkit.log.grpc.reporter.server_port`, and `plugin.toolkit.log.grpc.reporter.upstream_timeout`.
     rename `plugin.toolkit.log.grpc.reporter.max_message_size` to `log.max_message_size`.